% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimo.R
\name{nimo}
\alias{nimo}
\title{App for Species Distribution Modeling}
\usage{
nimo()
}
\description{
Entry function to launch nimo app to query occurrence data
from the Global Biodiversity Information Facility \href{https://www.gbif.org/occurrence/search}{(GBIF)}
and incorporate it into ecological niche modelling workflow offered by
\href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.13874}{flexdsm}
}
