% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_eval.R
\name{nm_eval}
\alias{nm_eval}
\title{Calculate model performance metrics}
\usage{
nm_eval(p, a, bg = NULL, thr = NULL)
}
\arguments{
\item{p}{numeric. Predicted suitability for presences}

\item{a}{numeric. Predicted suitability for absences}

\item{bg}{numeric. Predicted suitability for background points, used for BOYCE metric}

\item{thr}{character. Threshold criterion used to get binary suitability values (i.e. 0,1). See \emph{Details} section.}
}
\value{
A tibble with next columns
\itemize{
\item threshold: threshold names
\item thr_value: threshold values
\item n_presences: number of presences
\item n_absences: number of absences
\item performance metrics: from Prevalence to IMAE:
}
}
\description{
This function calculates threshold dependent and independent model performance metrics.
The function is imported and completed from \code{\link[=sdm_eval]{sdm_eval()}}.
}
\details{
The additionally metrics completed are: prevalence, False Positive Rate, False Nagative Rate,
Correct Classification Rate, Miss Classification Rate, Kappa. See \code{\link[=sdm_eval]{sdm_eval()}}.

It is possible to use more than one threshold type.
A vector must be provided for this argument. The following threshold criteria are available:
\itemize{
\item lpt: The highest threshold at which there is no omission.
\item equal_sens_spec: Threshold at which the Sensitivity and Specificity are equal.
\item max_sens_spec: Threshold at which the sum of the Sensitivity and Specificity
is the highest (aka threshold that maximizes the TSS).
\item max_jaccard: The threshold at which the Jaccard index is the highest.
\item max_sorensen: The threshold at which the Sorensen index is the highest.
\item max_fpb: The threshold at which FPB (F-measure on presence-background data) is the highest.
\item sensitivity: Threshold based on a specified Sensitivity value.
Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers
to Sensitivity value. If a sensitivity value is not specified, the
default value is 0.9
}
If more than one threshold type is used, concatenate threshold types,
e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec',
'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity').
Function will use all thresholds if no threshold type is specified
}
\examples{
\dontrun{
require(dplyr)

set.seed(0)
p <- rnorm(66, mean = 0.6, sd = 0.4) \%>\% abs()
p[p > 1] <- 1
p[p < 0] <- 0

set.seed(0)
a <- rnorm(45, mean = 0.6, sd = 0.3) \%>\% abs()
a[a > 1] <- 1
a[a < 0] <- 0

e <- nm_eval(p, a)
}

}
