% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_extract.R
\name{nm_extract}
\alias{nm_extract}
\title{Extract raster data values based on point coordinates (longitude, latitude)}
\usage{
nm_extract(
  data,
  longitude,
  latitude,
  env_layer,
  current_crs = "geographic",
  variables = NULL,
  filter_na = TRUE
)
}
\arguments{
\item{data}{data.frame. Database with species presence, presence-absence,
or pseudo-absence records with longitude and latitude coordinates.}

\item{longitude}{character. Column name with spatial longitude coordinates.}

\item{latitude}{character. Column name with spatial latitude coordinates.}

\item{env_layer}{SpatRaster. Raster or raster stack.}

\item{current_crs}{character. Current coordinate reference system of coordinates. One of:
\itemize{
\item \code{geographic}: Apllicable for longitude and latitude in decimal
(e.g 31.796406, -24.536200)
\item \code{projected}: Applicable for longitude and latitude in projected
system (e.g 7280635, 982148 - UTM zone 35S)
}}

\item{variables}{character. Vector with the raster names to extract value from.
If NULL (default) the function will return data for all raster.}

\item{filter_na}{logical. If filter_na = TRUE (default), the rows with NA values
for any of the raster are removed from the returned tibble.}
}
\value{
tibble with original data bound with extracted data
}
\description{
Extract raster data values based on point coordinates (longitude, latitude)
}
\examples{
 library(terra)
 wbv_path <- paste0(system.file("extdata", package = "nimo"), "/WbV_subset_occ_KNP.csv")
 wbv_df <- read.csv(wbv_path)

 env_layers_path <- paste0(system.file("extdata", package = "nimo"), "/env_layers")
 env_layers <- terra::rast(list.files(env_layers_path, full.names = TRUE)[c(1,3)])

extracted <- nm_extract(data = wbv_df, longitude = "decimalLongitude",
latitude = "decimalLatitude", env_layer = env_layers)

}
