% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_find_hcv.R
\name{nm_find_hcv}
\alias{nm_find_hcv}
\title{Find highly correlated variables}
\usage{
nm_find_hcv(x, cutoff = 0.9, verbose = FALSE)
}
\arguments{
\item{x}{A correlation matrix}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff}

\item{verbose}{A boolean for printing the details}
}
\value{
A vector of of column names denoting the columns to remove. If no correlations meet the criteria,
integer(0) is returned.
}
\description{
This function examines a correlation matrix and generates a vector of
integers representing columns to be eliminated for minimizing pairwise
correlations. It is similar to findCorrelation() in \code{caret} package.
}
\examples{
wbv_path <- paste0(system.file("extdata", package = "nimo"), "/Gyps_africanus_KNP.csv")
wbv_df <- read.csv(wbv_path)
mat_dt <- wbv_df[, c("dsm", "lst", "ndvi", "dnw")]
corr <- nm_find_hcv(x = cor(mat_dt), cutoff = 0.4)
corr

}
