% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_fit_bioclim.R
\name{nm_fit_bioclim}
\alias{nm_fit_bioclim}
\title{Bioclim model}
\usage{
nm_fit_bioclim(data, response, predictors, partition, thr = NULL)
}
\arguments{
\item{data}{data.frame object with response (0,1) and predictors values.}

\item{response}{character. Column name with species absence-presence data (0,1).}

\item{predictors}{character. Vector with the column names of quantitative predictor variables.}

\item{partition}{character. Column name with training and validation partition
groups (like part_* family output from \code{flexsdm}).}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1).
See a function of model fitting family in \code{flexsdm} package (e.g \code{\link[=fit_gam]{fit_gam()}})}

\item{...}{Additional arguments}
}
\value{
A list object with:
\itemize{
\item model: A "Bioclim" class object from dismo package. This object can be used for predicting.
\item predictors: A tibble with quantitative (c column names) variables use for modeling.
\item performance: Performance metric (see \code{\link{nm_eval}}).
Threshold dependent metrics are calculated based on the threshold specified in the argument.
\item data_ens: Predicted suitability for each test partition.
}
}
\description{
The BIOCLIM algorithm
}
\examples{

wbv_path <- paste0(system.file("extdata", package = "nimo"), "/Gyps_africanus_KNP.csv")
wbv_df <- read.csv(wbv_path)
mod = nm_fit_bioclim(data = wbv_df,
                    response = "pr_ab",
                    predictors = c("dnw", "dsm", "lst", "ndvi"),
                    partition = ".part",
                    thr = "max_sens_spec")

}
