% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_gbif_suggestion.R
\name{nm_gbif_suggestion}
\alias{nm_gbif_suggestion}
\title{Get suggestions from GBIF}
\usage{
nm_gbif_suggestion(query, time_out = 60)
}
\arguments{
\item{query}{Simple full text search parameter. The value for this parameter can be a simple word or a phrase. Wildcards are not supported}

\item{time_out}{Set time (in second) limit for a request. An error will be thrown
if the request does not complete in the time limit.}
}
\value{
Dataframe
}
\description{
A quick and simple autocomplete service that returns up to 20 name usages by doing
prefix matching against the scientific name. Results are ordered by relevance.
}
\examples{
nm_gbif_suggestion("Gyps")
}
