% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_predict.R
\name{nm_predict}
\alias{nm_predict}
\title{Spatial predictions from individual and ensemble models}
\usage{
nm_predict(
  models,
  pred,
  nchunk = 1,
  thr = NULL,
  con_thr = FALSE,
  predict_area = NULL,
  clamp = TRUE,
  pred_type = "cloglog"
)
}
\arguments{
\item{models}{list of one or more models fitted with fit_ or tune_ functions. In case use models fitted with fit_ensemble or esm_ family function only one model could be used. Usage models = mglm or models = list(mglm, mraf, mgbm)}

\item{pred}{SpatRaster. Raster layer with predictor variables. Names of layers must exactly
match those used in model fitting.}

\item{nchunk}{interger. Number of chunks to split data used to predict models (i.e., SpatRaster
used in pred argument). Predicting models in chunks helps reduce memory requirements in cases
where models are predicted for large scales and high resolution. Default = 1}

\item{thr}{character. Threshold used to get binary suitability values (i.e., 0,1). It is possible
to use more than one threshold type. It is mandatory to use the same threshold/s used to fit the
models. The following threshold types are available:
\itemize{
\item lpt: The highest threshold at which there is no omission.
\item equal_sens_spec: Threshold at which the sensitivity and specificity are equal.
\item max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the
highest (aka threshold that maximizes the TSS).
\item max_jaccard: The threshold at which the Jaccard index is the highest.
\item max_sorensen: The threshold at which the Sorensen index is highest.
\item max_fpb: The threshold at which FPB is highest.
\item sensitivity: Threshold based on a specified sensitivity value used to fit the models.
\item all: All the threshold used in the model outputs used in 'models' argument will be used.
}
Usage thr = c('lpt', 'max_sens_spec', 'max_jaccard'), thr=c('lpt', 'max_sens_spec',
'sensitivity'), or thr='all'. If no threshold is specified (i.e., thr = NULL) function
will return continuous prediction only. Default NULL}

\item{con_thr}{logical. If true predictions with suitability values above threshold/s will be
returned. Default = FALSE}

\item{predict_area}{SpatVector, SpatialPolygon, or SpatialPolygonDataFrame. Spatial polygon
used for restring prediction into only a given region. Default = NULL}

\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range
seen during model training. Only valid for Maxent model (see tune_mx and fit_mx). Default TRUE.}

\item{pred_type}{character. Type of response required available "link", "exponential",
"cloglog" and "logistic". Only valid for Maxent model (see tune_mx and fit_mx).
Default "cloglog".}
}
\value{
A list of SpatRaster with continuous and/or binary predictions
}
\description{
The function was initially designed by Velazco et al. (2022).
It is modified to extend the prediction on other class of model (e.g. bioclim).
The follow description is for initial authors.
This function allows the geographical prediction of one or more models constructed
with the fit_ or tune_ function set, models fitted with esm_ function set (i.e., ensemble of
small models approach), or models constructed with fit_ensemble function. It can return
continuous or continuous and binary predictions for one or more thresholds
}
\seealso{
\code{\link{sdm_predict}}
}
\author{
Velazco, S. J. E., Rose, M. B., de Andrade, A. F. A., Minoli, I., & Franklin, J. (2022).
flexsdm: An r package for supporting a comprehensive and flexible species distribution
modelling workflow. Methods in Ecology and Evolution, 13(8), 1661–1669.
\href{https://doi.org/10.1111/2041-210X.13874}{{https://doi.org/10.1111/2041-210X.13874}}
}
